+++ docs/conf.py
@@ -60,7 +60,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'mosaik'
-copyright = '2012-2021 OFFIS'
+copyright = '2012-2022 OFFIS'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the