+++ docs/tutorials/code/controller_master.py
@@ -12,7 +12,7 @@ META = {
         'Agent': {
             'public': True,
             'params': [],
-            'attrs': ['val_in', 'delta'],
+            'attrs': ['delta_in', 'delta_out'],
         },
     },
 }
@@ -23,13 +23,14 @@ class Controller(mosaik_api.Simulator):
         super().__init__(META)
         self.agents = []
         self.data = {}
+        self.cache = {}
         self.time = 0
 
     def create(self, num, model):
         n_agents = len(self.agents)
         entities = []
         for i in range(n_agents, n_agents + num):
-            eid = 'Agent_%d' % i
+            eid = 'Master_Agent_%d' % i
             self.agents.append(eid)
             entities.append({'eid': eid, 'type': model})
 
@@ -39,26 +40,12 @@ class Controller(mosaik_api.Simulator):
         self.time = time
         data = {}
         for agent_eid, attrs in inputs.items():
-            delta_dict = attrs.get('delta', {})
-            if len(delta_dict) > 0:
-                data[agent_eid] = {'delta': list(delta_dict.values())[0]}
-                continue
-
-            values_dict = attrs.get('val_in', {})
-            if len(values_dict) != 1:
-                raise RuntimeError('Only one ingoing connection allowed per '
-                                   'agent, but "%s" has %i.'
-                                   % (agent_eid, len(values_dict)))
-            value = list(values_dict.values())[0]
-
-            if value >= 3:
-                delta = -1
-            elif value <= -3:
-                delta = 1
-            else:
-                continue
-
-            data[agent_eid] = {'delta': delta}
+            values_dict = attrs.get('delta_in', {})
+            for key, value in values_dict.items():
+                self.cache[key] = value
+        
+        if sum(self.cache.values()) < -1:
+            data[agent_eid] = {'delta_out': 0}
 
         self.data = data
 
@@ -68,7 +55,7 @@ class Controller(mosaik_api.Simulator):
         data = {}
         for agent_eid, attrs in outputs.items():
             for attr in attrs:
-                if attr != 'delta':
+                if attr != 'delta_out':
                     raise ValueError('Unknown output attribute "%s"' % attr)
                 if agent_eid in self.data:
                     data['time'] = self.time